Compute type of inherited constructor correctly
Fixes #1246.
